+++ src/com/lld/inmemorycache/service/AbstractCache.java
@@ -1,8 +1,11 @@
-package com.lld.inmemorycache.model;
+package com.lld.inmemorycache.service;
+
+import com.lld.inmemorycache.service.EvictionPolicy;
+import com.lld.inmemorycache.service.Storage;
 
 public abstract class AbstractCache {
-    public IStorage storage;
-    public IEvictionPolicy evictionPolicy;
+    public Storage storage;
+    public EvictionPolicy evictionPolicy;
     public int capacity;
     public abstract boolean put(String key, String value);
     public abstract String get(String key);
