+++ src/com/lld/inmemorycache/service/EvictionPolicy.java
@@ -1,6 +1,6 @@
-package com.lld.inmemorycache.model;
+package com.lld.inmemorycache.service;
 
-public interface IEvictionPolicy {
+public interface EvictionPolicy {
     // Update statistics for the key which was accessed.
     public void keyAccessed(String key);
     // Update statistics for the key which was evicted.
