+++ src/com/lld/inmemorycache/service/Storage.java
@@ -1,6 +1,6 @@
-package com.lld.inmemorycache.model;
+package com.lld.inmemorycache.service;
 
-public interface IStorage {
+public interface Storage {
     public boolean put(String key, String value);
     public String get(String key);
     public boolean remove(String key);
